Move js/apps to /apps
