+++ cfgov/unprocessed/apps/financial-well-being/js/fwb-questions.js
@@ -1,4 +1,4 @@
-const Analytics = require( '../../modules/Analytics' );
+const Analytics = require( '../../../js/modules/Analytics' );
 
 /**
  * Initialize the questionnaire
