+++ cfgov/unprocessed/js/routes/consumer-tools/financial-well-being/index.js
@@ -1,4 +1,6 @@
-const fwbQuestions = require( '../../../apps/financial-well-being/fwb-questions' );
+const fwbQuestions = require(
+  '../../../../apps/financial-well-being/js/fwb-questions'
+);
 
 window.addEventListener( 'load', function() {
   fwbQuestions.init();
