+++ cfgov/unprocessed/js/routes/consumer-tools/financial-well-being/results/index.js
@@ -1,3 +1,5 @@
-const fwbResults = require( '../../../../apps/financial-well-being/fwb-results' );
+const fwbResults = require(
+  '../../../../../apps/financial-well-being/js/fwb-results'
+);
 
 fwbResults.init();
