+++ test/unit_tests/apps/financial-well-being/fwb-results-spec.js
@@ -1,4 +1,4 @@
-const BASE_JS_PATH = '../../../../cfgov/unprocessed/js/';
+const BASE_JS_PATH = '../../../../cfgov/unprocessed/apps/';
 const chai = require( 'chai' );
 const expect = chai.expect;
 const sinon = require( 'sinon' );
@@ -109,7 +109,7 @@ function triggerClickEvent( target ) {
 
 function initFwbResults( ) {
   fwbResults = require(
-    BASE_JS_PATH + 'apps/financial-well-being/fwb-results'
+    BASE_JS_PATH + 'financial-well-being/js/fwb-results'
   );
   fwbResults.init();
 }
